+++ CHANGELOG.md
@@ -139,3 +139,5 @@ but in the last case alias may be defined in runtime.
 - Move slug map item to array admin form data transformer to Content bundle.
 
 - Remove unsupported ar, fr, zh translations.
+
+- Rename "item factory pool" => "item factory registry".